% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{aov_all_vars}
\alias{aov_all_vars}
\title{Analysis of variance}
\usage{
aov_all_vars(
  dataset,
  column.class,
  doTukey = TRUE,
  write.file = FALSE,
  file.out = NULL
)
}
\arguments{
\item{dataset}{Dataset to analyze.}

\item{column.class}{Metadata column used to define groups.}

\item{doTukey}{Logical. If TRUE, also performs TukeyHSD post-hoc test.}

\item{write.file}{Logical. If TRUE, writes results to file.}

\item{file.out}{Output file name.}
}
\value{
A data frame with ANOVA results for all variables.
}
\description{
Performs one-way ANOVA for all variables in a dataset.
}
\examples{
datamatrix <- matrix(
  c(10, 11, 20, 21,
    5,  6,  5,  6),
  nrow = 2,
  byrow = TRUE,
  dimnames = list(c("x1", "x2"), c("s1", "s2", "s3", "s4"))
)
metadata <- data.frame(
  group = factor(c("A", "A", "B", "B")),
  row.names = c("s1", "s2", "s3", "s4")
)
dataset <- list(data = datamatrix, metadata = metadata)
aov_all_vars(dataset, "group", doTukey = FALSE)

}
\keyword{anova}
\keyword{tukey}
