% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{apply_by_group}
\alias{apply_by_group}
\title{Apply by group}
\usage{
apply_by_group(dataset, fn.to.apply, metadata.var, var.value)
}
\arguments{
\item{dataset}{Dataset to analyze.}

\item{fn.to.apply}{Function to apply.}

\item{metadata.var}{Metadata variable used to define the group.}

\item{var.value}{Value or values of the metadata variable to select.}
}
\value{
A named vector or matrix, depending on the output of `fn.to.apply`,
  containing the values obtained by applying the function to each variable
  across the samples belonging to the selected group.
}
\description{
Applies a function to the variables of samples belonging to a given group.
}
\examples{
data <- matrix(
  c(1, 2, 3, 4, 5, 6),
  nrow = 2,
  dimnames = list(c("x1", "x2"), c("s1", "s2", "s3"))
)
metadata <- data.frame(
  group = c("control", "control", "case"),
  row.names = c("s1", "s2", "s3")
)
dataset <- list(data = data, metadata = metadata)
apply_by_group(dataset, mean, "group", "control")

}
\keyword{apply}
\keyword{group}
