\docType{data}
\name{cachexia}
\alias{cachexia}
\title{Human Cachexia data}
\format{An object of class \code{"list"}}
\source{
\href{http://www.metaboanalyst.ca/resources/data/human_cachexia.csv}{MetaboAnalyst}
}
\usage{
data(cachexia)
}
\description{
Cachexia is a complex metabolic syndrome associated with an 
underlying illness (such as cancer) and characterized by loss of muscle 
with or without loss of fat mass (Evans et al., 2008).
A total of 77 urine samples were collected being 47 of them patients 
with cachexia, and 30 control patients.
}
\examples{
	data(cachexia)
    sum_dataset(cachexia)
}
\references{
Eisner et al. (2010) Learning to predict cancer-associated skeletal 
muscle wasting from 1h-nmr profiles of urinary metabolites 
Metabolomics 7:25-34
}
\keyword{datasets}
