\name{check_2d_dataset}
\alias{check_2d_dataset}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Check 2D dataset.
}
\description{
  Check if the dataset is valid and if not give the proper error message.
}
\usage{
check_2d_dataset(dataset_2d)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dataset_2d}{
  List representing the 2D dataset from a 2D metabolomics experiment.
}
}

\value{
  Message saying if the 2D dataset is valid or invalid, in the last case also gives the reason.
}




% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
\keyword{ 2d-dataset }

