\name{conversion_table}
\alias{conversion_table}
\docType{data}
\title{
Conversion table.
}
\description{
Conversion table is a dataframe with information on the KEGG code, common name, 1D and 2D-NMR reference spectra ID and chebi code for each HMDB metabolite entry.
}
\usage{data("conversion_table")}
