% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHWAY_ANALYSIS.R
\name{convert_chebi_to_kegg}
\alias{convert_chebi_to_kegg}
\title{Get kegg codes from chebi codes:}
\usage{
convert_chebi_to_kegg(chebi_codes)
}
\arguments{
\item{chebi_codes}{Character vector of ChEBI identifiers.}
}
\value{
A named \code{character} vector with the KEGG compound identifiers
  corresponding to the input ChEBI codes. The vector values are KEGG compound
  codes prefixed with \code{"cpd:"}, and the element names are the matched
  compound names.
}
\description{
Auto-exported function: convert_chebi_to_kegg
}
\examples{
\dontrun{
convert_chebi_to_kegg(c("CHEBI:15377"))
}

}
\keyword{internal}
