% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHWAY_ANALYSIS.R
\name{convert_hmdb_to_kegg}
\alias{convert_hmdb_to_kegg}
\title{Get kegg codes from hmdb codes:}
\usage{
convert_hmdb_to_kegg(hmdb_codes)
}
\arguments{
\item{hmdb_codes}{Character vector of HMDB identifiers.}
}
\value{
A named \code{character} vector with the KEGG compound identifiers
  corresponding to the input HMDB codes. The vector values are KEGG compound
  codes prefixed with \code{"cpd:"}, and the element names are the matched
  compound names.
}
\description{
Get kegg codes from hmdb codes:
}
\examples{
\dontrun{
convert_hmdb_to_kegg(c("HMDB0000122"))
}
}
