\name{count.missing.values}
\alias{count.missing.values}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Count missing values
}
\description{
  Counts the missing values on the dataset.
}
\usage{
count.missing.values(dataset)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dataset}{
    list representing the dataset from a metabolomics experiment.
}
}

\value{
  Returns the number of missing values on the dataset.
}
\examples{
\dontrun{
  ## Example of counting the missing values
  count.missing.values(dataset)
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~missing }
\keyword{ ~values }% __ONLY ONE__ keyword per line
