\name{cubic.root.transform}
\alias{cubic.root.transform}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Cubic root transformation
}
\description{
  Performs cubic root transformation on the data matrix.
}
\usage{
cubic.root.transform(datamat)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{datamat}{
   data matrix.
}
}

\value{
   Returns the data matrix with the cubic root transformation applied.
}

\examples{
\dontrun{
  ## Example of cubic root transformation
  datamat.cubic = cubic.root.transform(datamatrix)
}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~cubicroot }
\keyword{ ~transformation }% __ONLY ONE__ keyword per line
