% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{filter_feature_selection}
\alias{filter_feature_selection}
\title{Feature Selection Using Univariate Filters}
\usage{
filter_feature_selection(
  datamat,
  samples.class,
  functions = caret::rfSBF,
  method = "cv",
  repeats = 5
)
}
\arguments{
\item{datamat}{Data matrix with features in rows and samples in columns.}

\item{samples.class}{Sample class labels.}

\item{functions}{Caret SBF functions list.}

\item{method}{Resampling method passed to \code{caret::sbfControl()}.}

\item{repeats}{Number of repeats for resampling.}
}
\value{
An \code{sbf} object.
}
\description{
Performs feature selection using univariate filters.
}
\examples{
datamat <- matrix(
  rnorm(20),
  nrow = 4,
  dimnames = list(paste0("x", 1:4), paste0("s", 1:5))
)
classes <- factor(c("A", "A", "B", "B", "A"))
filter_feature_selection(datamat, classes, caret::rfSBF)
}
