\name{get.x.label}
\alias{get.x.label}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Get x-axis label
}
\description{
  Get the x-axis label from the dataset.
}
\usage{
get.x.label(dataset)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dataset}{
    list representing the dataset from a metabolomics experiment.
}
}

\value{
  Returns a string with the x-axis label.
}

\examples{
\dontrun{
  ## Example of getting the x-axis label
  x.label = get.x.label(dataset)
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~label }
\keyword{ ~xaxis }% __ONLY ONE__ keyword per line
