% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHWAY_ANALYSIS.R
\name{get_cpd_names}
\alias{get_cpd_names}
\title{Get compound names from KEGG codes}
\usage{
get_cpd_names(kegg_codes)
}
\arguments{
\item{kegg_codes}{Character vector of KEGG compound identifiers.}
}
\value{
A named \code{character} vector in which the values are KEGG compound
  identifiers and the names are the corresponding compound names. Each element
  of the returned vector represents one KEGG compound code matched to its
  primary compound name.
}
\description{
Returns compound names associated with KEGG compound identifiers.
}
\examples{
\dontrun{
get_cpd_names(c("C00031", "C00022"))
}
}
