% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metabolights_study.R
\name{get_metabolights_study_files_assay}
\alias{get_metabolights_study_files_assay}
\title{Download files for one MetaboLights assay}
\usage{
get_metabolights_study_files_assay(studyID, assay, directory, verbose = TRUE)
}
\arguments{
\item{studyID}{Character string with the MetaboLights study identifier.}

\item{assay}{Numeric or character index selecting the assay in the study.}

\item{directory}{Output directory where assay files will be downloaded.}

\item{verbose}{Logical indicating whether progress messages should be shown.}
}
\value{
Invisibly returns a \code{character} vector with downloaded file paths.
}
\description{
Downloads the data files associated with one assay of a MetaboLights study.
}
\examples{
\dontrun{
tmpdir <- tempdir()
get_metabolights_study_files_assay("MTBLS1", assay = 1, directory = tmpdir, verbose = FALSE)
}

}
