% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{get_x_values_as_text}
\alias{get_x_values_as_text}
\title{Get x values as text}
\usage{
get_x_values_as_text(dataset)
}
\arguments{
\item{dataset}{Dataset object.}
}
\value{
A character vector containing the variable identifiers stored in
  `rownames(dataset$data)`. Each element corresponds to one row of the data
  matrix and represents the x-axis value or variable label associated with
  that feature.
}
\description{
Returns the x values of a dataset as text.
}
\examples{
datamatrix <- matrix(
  c(1, 2, 3, 4),
  nrow = 2,
  dimnames = list(c("10.5", "11.0"), c("s1", "s2"))
)
metadata <- data.frame(class = c("A", "B"), row.names = c("s1", "s2"))
dataset <- create_dataset(
  datamatrix,
  type = "concentrations",
  metadata = metadata
)
get_x_values_as_text(dataset)

}
