% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICA.R
\name{ica_kmeans_plot2D}
\alias{ica_kmeans_plot2D}
\title{ICA 2D k-means plot}
\usage{
ica_kmeans_plot2D(
  dataset,
  ica.result,
  num.clusters = 3,
  dims = c(1, 2),
  kmeans.result = NULL,
  use.embedding = TRUE,
  labels = FALSE,
  bw = FALSE,
  ellipses = FALSE,
  leg.pos = "right",
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{dataset}{Dataset to cluster.}

\item{ica.result}{Result returned by \code{ica_analysis_dataset()}.}

\item{num.clusters}{Number of k-means clusters.}

\item{dims}{Two independent components to plot.}

\item{kmeans.result}{Optional k-means result containing a \code{cluster}
vector.}

\item{use.embedding}{Logical indicating whether k-means should be applied
to the ICA embedding when \code{kmeans.result} is not supplied.}

\item{labels}{Logical indicating whether sample labels should be displayed.}

\item{bw}{Logical indicating whether a black-and-white plot should be used.}

\item{ellipses}{Logical indicating whether group ellipses should be
displayed.}

\item{leg.pos}{Legend position.}

\item{xlim}{Optional x-axis limits.}

\item{ylim}{Optional y-axis limits.}
}
\value{
A \code{ggplot} object.
}
\description{
Creates a two-dimensional ICA plot coloured by k-means clusters.
}
\examples{
if (requireNamespace("fastICA", quietly = TRUE)) {
  datamat <- matrix(rnorm(240), nrow = 8)
  dataset <- list(
    data = datamat,
    metadata = data.frame(
      class = factor(rep(c("A", "B"), each = 15))
    )
  )

  ica.result <- ica_analysis_dataset(
    dataset,
    n_components = 2,
    maxit = 100,
    seed = 42
  )

  kmeans.result <- list(
    cluster = rep(1:2, each = 15)
  )

  ica_kmeans_plot2D(
    dataset,
    ica.result,
    num.clusters = 2,
    kmeans.result = kmeans.result
  )
}

}
