% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{impute_nas_value}
\alias{impute_nas_value}
\title{Impute missing values with a constant}
\usage{
impute_nas_value(dataset, value)
}
\arguments{
\item{dataset}{A dataset object to modify.}

\item{value}{A numeric or character value used to replace missing entries.}
}
\value{
A dataset object with the same structure as the input, where all
  missing values in `dataset$data` have been replaced by `value`. Other
  components of the dataset are preserved unchanged.
}
\description{
Replace all `NA` values in `dataset$data` by a user-defined constant.
}
\examples{
data <- matrix(
  c(1, NA, 3, 4),
  nrow = 2,
  dimnames = list(c("x1", "x2"), c("s1", "s2"))
)
dataset <- list(data = data)
impute_nas_value(dataset, 0)

}
