% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMR_metabolite_identification.R
\name{jaccard_index}
\alias{jaccard_index}
\title{JACCARD INDEX}
\usage{
jaccard_index(peaksCluster, peaksReference, PPMTOL)
}
\description{
Calculates the jaccard index, i.e., the similarity between cluster and reference metabolite.
}
