% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metabolights_study.R
\name{metabolights_studies_list}
\alias{metabolights_studies_list}
\title{List public MetaboLights studies}
\usage{
metabolights_studies_list()
}
\value{
A \code{character} vector with public MetaboLights study identifiers.
}
\description{
Returns the identifiers of public studies available in MetaboLights.
}
\examples{
\dontrun{
head(metabolights_studies_list())
}

}
