\name{msc_correction}
\alias{msc_correction}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Multiplicative scatter correction
}
\description{
  Perform multiplicative scatter correction on the spectra.
}
\usage{
msc_correction(dataset)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dataset}{
    list representing the dataset from a metabolomics experiment.
}
}

\value{
  Return the dataset with the multiplicative scatter correction employed on the data.
}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ msc }
\keyword{ dataset }% __ONLY ONE__ keyword per line
