\name{nmr_1d_spectra_options}
\alias{nmr_1d_spectra_options}
\docType{data}
\title{
Conditions of acquirement of the different refernece peak lists.
}
\description{
Dataframe with what conditions each reference peak list was obtained: frequency, nuleus, pH, temperature and solvent. This dataset is used to filter the reference metabolites in the identification of 1D-NMR peak lists samples.
}
\usage{data("nmr_1d_spectra_options")}
