% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMR_metabolite_identification.R
\name{nmr_clustering}
\alias{nmr_clustering}
\title{CLUSTER OF VARIABLES (PEAKS)}
\usage{
nmr_clustering(dataset_orig, CMETH = "pearson", CVAL = 0.95, MVER = 2)
}
\arguments{
\item{dataset_orig}{List representing the dataset from an nmr peaks metabolomics experiment.}

\item{CMETH}{Correlation method used to cluster the variables. Defaults to "pearson".}

\item{CVAL}{Minimum correlation between variables, so they can belong to the same cluster.}

\item{MVER}{Minimum number of variables in each cluster. Only the clusters with at least MVER variables will be returned.}
}
\value{
List with the formed clusters
}
\description{
Takes a dataset and performs clustering of variables, according to a correlation. The variables
will be separated into different clusters, according to a minimum correlation between variables.
Each cluster will correspond to a metabolite.
}
