% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_biplot}
\alias{pca_biplot}
\title{PCA biplot}
\usage{
pca_biplot(
  dataset,
  pca.result,
  cex = 0.8,
  legend.cex = 0.8,
  x.colors = 1,
  inset = c(0, 0),
  legend.place = "topright",
  ...
)
}
\arguments{
\item{dataset}{Dataset used in the PCA.}

\item{pca.result}{PCA result object.}

\item{cex}{Text expansion factor.}

\item{legend.cex}{Legend text size.}

\item{x.colors}{Colour vector or metadata column name.}

\item{inset}{Legend inset.}

\item{legend.place}{Legend position.}

\item{...}{Additional arguments passed to the plotting method.}
}
\value{
No return value, called for its side effect of plotting.
}
\description{
Draws a PCA biplot for a \code{prcomp} or \code{princomp} result.
}
\examples{
datamat <- matrix(
  rnorm(20),
  nrow = 4,
  dimnames = list(paste0("x", 1:4), paste0("s", 1:5))
)
metadata <- data.frame(class = factor(c("A", "A", "B", "B", "A")))
dataset <- list(data = datamat, metadata = metadata)
pca_res <- pca_analysis_dataset(dataset)
pca_biplot(dataset, pca_res)

}
