% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_biplot3D}
\alias{pca_biplot3D}
\title{PCA 3D biplot}
\usage{
pca_biplot3D(dataset, pca.result, column.class = NULL, pcas = c(1, 2, 3))
}
\arguments{
\item{dataset}{Dataset used in the PCA.}

\item{pca.result}{PCA result object.}

\item{column.class}{Optional metadata column used for colouring groups.}

\item{pcas}{Principal components to plot.}
}
\value{
No return value, called for its side effect of plotting.
}
\description{
Creates an interactive 3D PCA biplot using \code{rgl}.
}
\examples{
\donttest{
datamat <- matrix(
  rnorm(20),
  nrow = 4,
  dimnames = list(paste0("x", 1:4), paste0("s", 1:5))
)
metadata <- data.frame(class = factor(c("A", "A", "B", "B", "A")))
dataset <- list(data = datamat, metadata = metadata)
pca_res <- pca_analysis_dataset(dataset)
pca_biplot3D(dataset, pca_res, column.class = "class")
}

}
