% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_kmeans_plot2D}
\alias{pca_kmeans_plot2D}
\title{PCA 2D k-means plot}
\usage{
pca_kmeans_plot2D(
  dataset,
  pca.result,
  num.clusters = 3,
  pcas = c(1, 2),
  kmeans.result = NULL,
  labels = FALSE,
  bw = FALSE,
  ellipses = FALSE,
  leg.pos = "right",
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{dataset}{Dataset used in the PCA.}

\item{pca.result}{PCA result object.}

\item{num.clusters}{Number of clusters.}

\item{pcas}{Principal components to plot.}

\item{kmeans.result}{Optional precomputed k-means result.}

\item{labels}{Logical indicating whether sample labels should be shown.}

\item{bw}{Logical indicating whether a black-and-white style should be used.}

\item{ellipses}{Logical indicating whether cluster ellipses should be drawn.}

\item{leg.pos}{Legend position.}

\item{xlim}{Optional x-axis limits.}

\item{ylim}{Optional y-axis limits.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots PCA scores in 2D coloured by k-means cluster assignment.
}
\examples{
\donttest{
datamat <- matrix(
  rnorm(20),
  nrow = 4,
  dimnames = list(paste0("x", 1:4), paste0("s", 1:5))
)
metadata <- data.frame(class = factor(c("A", "A", "B", "B", "A")))
dataset <- list(data = datamat, metadata = metadata)
pca_res <- pca_analysis_dataset(dataset)
pca_kmeans_plot2D(dataset, pca_res, num.clusters = 2)
}

}
