% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_pairs_kmeans_plot}
\alias{pca_pairs_kmeans_plot}
\title{PCA pairs plot with k-means clusters}
\usage{
pca_pairs_kmeans_plot(
  dataset,
  pca.result,
  num.clusters = 3,
  kmeans.result = NULL,
  pcas = c(1, 2, 3, 4, 5)
)
}
\arguments{
\item{dataset}{Dataset used in the PCA.}

\item{pca.result}{PCA result object.}

\item{num.clusters}{Number of clusters.}

\item{kmeans.result}{Optional precomputed k-means result.}

\item{pcas}{Principal components to include.}
}
\value{
A \code{ggmatrix} object.
}
\description{
Creates a pairs plot of PCA scores coloured by k-means cluster assignment.
}
\examples{
\donttest{
datamat <- matrix(
  rnorm(30),
  nrow = 6,
  dimnames = list(paste0("x", 1:6), paste0("s", 1:5))
)
metadata <- data.frame(class = factor(c("A", "A", "B", "B", "A")))
dataset <- list(data = datamat, metadata = metadata)
pca_res <- pca_analysis_dataset(dataset)
pca_pairs_kmeans_plot(dataset, pca_res, num.clusters = 2, pcas = 1:3)
}

}
