% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_screeplot}
\alias{pca_screeplot}
\title{PCA scree plot}
\usage{
pca_screeplot(
  pca.result,
  num.pcs = NULL,
  cex.leg = 0.8,
  leg.pos = "right",
  lab.text = c("individual percent", "cumulative percent"),
  fill.col = c("blue", "red"),
  ylab = "Percentage",
  xlab = "Principal components",
  ...
)
}
\arguments{
\item{pca.result}{PCA result object.}

\item{num.pcs}{Number of principal components to display.}

\item{cex.leg}{Legend text size.}

\item{leg.pos}{Legend position.}

\item{lab.text}{Legend labels.}

\item{fill.col}{Line colours.}

\item{ylab}{Y-axis label.}

\item{xlab}{X-axis label.}

\item{...}{Additional graphical parameters.}
}
\value{
No return value, called for its side effect of plotting.
}
\description{
Draws a scree plot with individual and cumulative explained variance.
}
\examples{
datamat <- matrix(
  rnorm(20),
  nrow = 4,
  dimnames = list(paste0("x", 1:4), paste0("s", 1:5))
)
metadata <- data.frame(class = factor(c("A", "A", "B", "B", "A")))
dataset <- list(data = datamat, metadata = metadata)
pca_res <- pca_analysis_dataset(dataset)
pca_screeplot(pca_res)

}
