\docType{data}
\name{propolisSampleList}
\alias{propolisSampleList}
\title{Brazilian Propolis from different Harvest Seasons and different Agroecological Regions (sample list)}
\format{An object of class \code{"list"}}

\usage{
data(propolisSampleList)
}
\description{
Propolis or bee glue is a sticky dark-colored substance produced 
from the collected buds or exudates of plants (resin) by bees 
(Apis mellifera L.). The resin is masticated, salivary enzymes are 
added, and the partially digested material is mixed with beewax 
and used in the hive to seal the walls, strengthen the borders of 
combs, and embalm dead invaders (Wollenweber et al., 1990).
The propolis samples are from NMR data and were collected in the 
autumn (AU), winter (WI), spring (SP), and summer (SM) of 2010 
from Apis mellifera hives located in southern Brazil (Santa 
Catarina State). A total of 59 samples were collected, and the 
distribution of samples by seasons being: SM - 16 samples, AU 
and SP - 15 samples, WI - 13 samples. Also, three agroecological 
regions were defined for the different apiaries, and one 
distributed as follows: Highlands - 12 samples, Plain - 11 samples, 
Plateau - 36 samples.
}
\examples{
    data(propolisSampleList)
    propolisSampleList[[1]]
}
\references{
E. Wollenweber, B. M. Hausen, and W. Greenaway. Phenolic 
constituents and sensitizing properties of propolis, poplar 
balsam and balsam of peru. Bulletin de Groupe Polyphenol, 
15:112-120, 1990.
M. Maraschin, A. Somensi-Zeggio, S. K. Oliveira, S. Kuhnen, 
M. M. Tomazzoli, A. C. M. Zeri, R. Carreira, and M. Rocha. 
A machine learning and chemometrics assisted interpretation 
of spectroscopic data - a nmr-based metabolomics platform for 
the assessment of brazilian propolis. 2012
}
\keyword{datasets}
