% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.spc.modifed.R
\name{read_spc_nosubhdr}
\alias{read_spc_nosubhdr}
\title{Import for Thermo Galactic's spc file format
These functions allow to import .spc files.
A detailed description of the .spc file format is available at}
\usage{
read_spc_nosubhdr(
  filename,
  keys.hdr2data = c("fexper", "fres", "fsource"),
  keys.hdr2log = c("fdate", "fpeakpt"),
  keys.log2data = FALSE,
  keys.log2log = TRUE,
  log.txt = TRUE,
  log.bin = FALSE,
  log.disk = FALSE,
  hdr = list(),
  nosubhdr = TRUE,
  no.object = FALSE
)
}
\arguments{
\item{filename}{The complete file name of the .spc file.}

\item{keys.hdr2data, keys.hdr2log, keys.log2data, keys.log2log}{character
vectors with the names of parameters in the .spc file's log block
(log2xxx) or header (hdr2xxx) that should go into the extra data
(yyy2data) or into the \code{long.description} field of the returned
object's log (yyy2log).}

\item{log.txt}{Should the text part of the .spc file's log block be read?}

\item{log.bin, log.disk}{Should the normal and on-disk binary parts of the
.spc file's log block be read?}

\item{hdr}{A list with fileheader fields that overwrite the settings of
actual file's header.}

\item{nosubhdr}{Boolean value to decide if the header should be read or not.}

\item{no.object}{If \code{TRUE}, a list with wavelengths, spectra, labels,
log and data are returned instead of a hyperSpec object.}
}
\value{
A list with imported spectra information.
}
\description{
Import for Thermo Galactic's spc file format
These functions allow to import .spc files.
A detailed description of the .spc file format is available at
}
\examples{
\donttest{
filename <- system.file("extdata", "toy.spc", package = "specmine")
if (nzchar(filename)) {
  read_spc_nosubhdr(filename)
}
}
}
\references{
Reference information for the SPC file format.
}
\author{
C. Beleites
}
