% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dx.R
\name{read_data_dx}
\alias{read_data_dx}
\title{Read JDX spectra files from a folder}
\usage{
read_data_dx(foldername, debug = 0, verbose = TRUE)
}
\arguments{
\item{foldername}{Path to the folder containing JDX files.}

\item{debug}{Debug level passed to \code{readJDX::readJDX()}.}

\item{verbose}{Logical; whether progress messages should be shown.}
}
\value{
A named list with one imported JDX object per file.
}
\description{
Reads all \code{.dx} files in a folder using \code{readJDX::readJDX()}.
}
\examples{
folder <- tempdir()
\dontrun{
read_data_dx(folder)
}

}
