% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{remove_data}
\alias{remove_data}
\title{Remove data}
\usage{
remove_data(
  dataset,
  data.to.remove,
  type = "sample",
  by.index = FALSE,
  rebuild.factors = TRUE
)
}
\arguments{
\item{dataset}{Dataset to modify.}

\item{data.to.remove}{Data to remove.}

\item{type}{One of "sample", "data", or "metadata".}

\item{by.index}{Logical. If TRUE, data.to.remove has indexes.}

\item{rebuild.factors}{Logical. If TRUE, rebuild factors in metadata.}
}
\value{
A dataset object with the same overall structure as the input, where
  the requested samples, data variables, or metadata variables have been
  removed according to `type`.
}
\description{
Removes selected samples, data variables, or metadata variables from a dataset.
}
\examples{
data <- matrix(
  1:9,
  nrow = 3,
  dimnames = list(c("x1", "x2", "x3"), c("s1", "s2", "s3"))
)
metadata <- data.frame(class = c("A", "B", "A"), row.names = c("s1", "s2", "s3"))
dataset <- list(data = data, metadata = metadata)
remove_data(dataset, "s2", type = "sample")

}
