% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{remove_data_variables}
\alias{remove_data_variables}
\title{Remove data variables}
\usage{
remove_data_variables(dataset, variables.to.remove, by.index = FALSE)
}
\arguments{
\item{dataset}{Dataset to modify.}

\item{variables.to.remove}{Variables to remove.}

\item{by.index}{Logical. If TRUE, variables.to.remove are indexes.}
}
\value{
A dataset object with the same structure as the input, where the
  selected rows of `dataset$data` have been removed. If no matching variables
  are found, the input dataset is returned unchanged and a warning is issued.
}
\description{
Removes x variables from a dataset.
}
\examples{
data <- matrix(
  1:9,
  nrow = 3,
  dimnames = list(c("10", "20", "30"), c("s1", "s2", "s3"))
)
metadata <- data.frame(class = c("A", "B", "A"), row.names = c("s1", "s2", "s3"))
dataset <- list(data = data, metadata = metadata)
remove_data_variables(dataset, "20")

}
