% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{remove_variables_by_nas}
\alias{remove_variables_by_nas}
\title{Remove variables by NAs}
\usage{
remove_variables_by_nas(dataset, max.nas = 0, by.percent = FALSE)
}
\arguments{
\item{dataset}{Dataset to modify.}

\item{max.nas}{Maximum number of missing values allowed.}

\item{by.percent}{Logical. If TRUE, max.nas is treated as a percentage.}
}
\value{
A dataset object with the same structure as the input, where variables
  whose number of missing values exceeds `max.nas` have been removed from
  `dataset$data`.
}
\description{
Removes variables with too many missing values.
}
\examples{
data <- matrix(
  c(1,  2, 3,
    NA, 5, NA,
    7,  8, 9),
  nrow = 3,
  dimnames = list(c("x1", "x2", "x3"), c("s1", "s2", "s3"))
)
metadata <- data.frame(class = c("A", "B", "A"), row.names = c("s1", "s2", "s3"))
dataset <- list(data = data, metadata = metadata)
remove_variables_by_nas(dataset, max.nas = 1)

}
