\docType{data}
\name{spinalCord}
\alias{spinalCord}
\title{Brazilian Propolis from different Harvest Seasons and different Agroecological Regions}
\format{An object of class \code{"list"}}
\source{
\href{http://www.metaboanalyst.ca/resources/data/lcms_netcdf.zip}{MetaboAnalyst}
}
\usage{
data(spinalCord)
}
\description{
This dataset consists of 12 LC-MS samples of spectra in the netCDF 
format, from mice spinal cord divided into 2 groups: the wild type 
and the knockout group. The data was obtained from the 
MetaboAnalyst site, originating from a study which describes a 
general strategy for identifying endogenous substrates of enzymes 
by untargeted LC-MS analysis of tissue metabolomes from
wild-type and enzyme-inactivated organisms
}
\examples{
    data(spinalCord)
    sum_dataset(spinalCord)
}
\references{
A. Saghatelian, S.A. Trauger, E.J. Want, E.G. Hawkins, G. Siuzdak,
B.F. Cravatt Assignment of Endogenous Substrates to Enzymes 
by Global Metabolite Profiling Biochemistry, 43:14332-14339, 2004.
}
\keyword{datasets}
