% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{subset_metadata}
\alias{subset_metadata}
\title{Subset metadata}
\usage{
subset_metadata(dataset, variables)
}
\arguments{
\item{dataset}{Dataset to subset.}

\item{variables}{Metadata variables to keep.}
}
\value{
A dataset object with the same structure as the input, where
  `dataset$metadata` contains only the selected metadata variables.
}
\description{
Selects metadata variables to keep.
}
\examples{
data <- matrix(
  1:6,
  nrow = 2,
  dimnames = list(c("x1", "x2"), c("s1", "s2", "s3"))
)
metadata <- data.frame(
  class = c("A", "B", "A"),
  batch = c(1, 1, 2),
  row.names = c("s1", "s2", "s3")
)
dataset <- list(data = data, metadata = metadata)
subset_metadata(dataset, "class")

}
