% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{subset_random_samples}
\alias{subset_random_samples}
\title{Subset random samples}
\usage{
subset_random_samples(dataset, nsamples)
}
\arguments{
\item{dataset}{Dataset to subset.}

\item{nsamples}{Number of samples to select.}
}
\value{
A dataset object with the same structure as the input, containing a
  random subset of `nsamples` samples.
}
\description{
Selects a random subset of samples from the dataset.
}
\examples{
set.seed(123)
data <- matrix(
  1:12,
  nrow = 3,
  dimnames = list(c("x1", "x2", "x3"), c("s1", "s2", "s3", "s4"))
)
metadata <- data.frame(
  class = factor(c("A", "A", "B", "B")),
  row.names = c("s1", "s2", "s3", "s4")
)
dataset <- list(data = data, metadata = metadata)
subset_random_samples(dataset, 2)

}
