% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{subset_samples}
\alias{subset_samples}
\title{Subset samples}
\usage{
subset_samples(dataset, samples, rebuild.factors = TRUE)
}
\arguments{
\item{dataset}{Dataset to subset.}

\item{samples}{Vector with indexes or names of the samples to select.}

\item{rebuild.factors}{If TRUE, rebuild factors in metadata.}
}
\value{
A dataset object with the same overall structure as the input, containing
  only the selected samples in both `dataset$data` and `dataset$metadata`. If
  `rebuild.factors` is `TRUE`, unused factor levels in metadata are dropped.
}
\description{
Returns a dataset with a selected set of samples.
}
\examples{
data <- matrix(
  c(1, 2, 3, 4, 5, 6),
  nrow = 2,
  dimnames = list(c("x1", "x2"), c("s1", "s2", "s3"))
)
metadata <- data.frame(
  class = factor(c("A", "B", "A")),
  row.names = c("s1", "s2", "s3")
)
dataset <- list(data = data, metadata = metadata)
subset_samples(dataset, c("s1", "s3"))

}
