% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{subset_samples_by_metadata_values}
\alias{subset_samples_by_metadata_values}
\title{Subset samples by metadata values}
\usage{
subset_samples_by_metadata_values(dataset, metadata.varname, values)
}
\arguments{
\item{dataset}{Dataset to subset.}

\item{metadata.varname}{Metadata variable name.}

\item{values}{Values to keep.}
}
\value{
A dataset object with the same structure as the input, containing only
  samples whose `metadata.varname` value matches one of the requested `values`.
}
\description{
Selects a set of samples by the value of a metadata variable.
}
\examples{
data <- matrix(
  c(1, 2, 3, 4, 5, 6),
  nrow = 2,
  dimnames = list(c("x1", "x2"), c("s1", "s2", "s3"))
)
metadata <- data.frame(
  class = factor(c("A", "B", "A")),
  row.names = c("s1", "s2", "s3")
)
dataset <- list(data = data, metadata = metadata)
subset_samples_by_metadata_values(dataset, "class", "A")

}
