% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{subset_x_values}
\alias{subset_x_values}
\title{Subset x values}
\usage{
subset_x_values(dataset, variables, by.index = FALSE)
}
\arguments{
\item{dataset}{Dataset to subset.}

\item{variables}{Variables to keep.}

\item{by.index}{Logical. If TRUE, variables are interpreted as row indexes.}
}
\value{
A dataset object with the same structure as the input, where
  `dataset$data` contains only the selected variables or row indexes.
}
\description{
Selects rows of the dataset by x values or by row index.
}
\examples{
data <- matrix(
  1:9,
  nrow = 3,
  dimnames = list(c("10", "20", "30"), c("s1", "s2", "s3"))
)
metadata <- data.frame(group = c("A", "B", "A"), row.names = c("s1", "s2", "s3"))
dataset <- list(data = data, metadata = metadata)
subset_x_values(dataset, c("10", "30"))

}
