% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{subset_x_values_by_interval}
\alias{subset_x_values_by_interval}
\title{Subset x values by interval}
\usage{
subset_x_values_by_interval(dataset, min.value, max.value)
}
\arguments{
\item{dataset}{Dataset to subset.}

\item{min.value}{Minimum x value.}

\item{max.value}{Maximum x value.}
}
\value{
A dataset object with the same structure as the input, where
  `dataset$data` contains only variables whose x values fall between
  `min.value` and `max.value`, inclusive.
}
\description{
Selects x values within a numeric interval.
}
\examples{
data <- matrix(
  1:12,
  nrow = 4,
  dimnames = list(c("10", "20", "30", "40"), c("s1", "s2", "s3"))
)
metadata <- data.frame(group = c("A", "B", "A"), row.names = c("s1", "s2", "s3"))
dataset <- list(data = data, metadata = metadata)
subset_x_values_by_interval(dataset, 15, 35)

}
