% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{umap_kmeans_plot2D}
\alias{umap_kmeans_plot2D}
\title{UMAP 2D k-means plot}
\usage{
umap_kmeans_plot2D(
  dataset,
  umap.result,
  num.clusters = 3,
  dims = c(1, 2),
  kmeans.result = NULL,
  labels = FALSE,
  bw = FALSE,
  ellipses = FALSE,
  leg.pos = "right",
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{dataset}{Dataset to cluster.}

\item{umap.result}{Result returned by \code{umap_analysis_dataset()}.}

\item{num.clusters}{Number of k-means clusters.}

\item{dims}{Two embedding dimensions to plot.}

\item{kmeans.result}{Optional result returned by \code{clustering()}.}

\item{labels}{Logical indicating whether sample labels should be displayed.}

\item{bw}{Logical indicating whether a black-and-white plot should be used.}

\item{ellipses}{Logical indicating whether group ellipses should be displayed.}

\item{leg.pos}{Legend position.}

\item{xlim}{Optional x-axis limits.}

\item{ylim}{Optional y-axis limits.}
}
\value{
A \code{ggplot} object.
}
\description{
Creates a two-dimensional UMAP plot coloured by k-means clusters.
}
\examples{
if (requireNamespace("uwot", quietly = TRUE)) {
  datamat <- matrix(
    rnorm(40),
    nrow = 5,
    dimnames = list(
      paste0("x", 1:5),
      paste0("s", 1:8)
    )
  )

  dataset <- list(
    data = datamat,
    metadata = data.frame(
      class = factor(rep(c("A", "B"), length.out = 8))
    )
  )

  umap.result <- umap_analysis_dataset(
    dataset,
    n_components = 2,
    n_neighbors = 3,
    seed = 42
  )

  umap_kmeans_plot2D(
    dataset,
    umap.result,
    num.clusters = 2,
    dims = c(1, 2)
  )
}

}
