% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{aggregate_samples}
\alias{aggregate_samples}
\title{Aggregate samples}
\usage{
aggregate_samples(dataset, indexes, aggreg.fn = "mean", meta.to.remove = c())
}
\arguments{
\item{dataset}{List representing the dataset from a metabolomics experiment.}

\item{indexes}{Index vector with the samples that are going to be aggregated
(e.g. c(1,1,2,2), this index vector will aggregate the first two samples and
the last two samples).}

\item{aggreg.fn}{Aggregation function (e.g. "mean", "median", etc).}

\item{meta.to.remove}{Metadata variables to be removed.}
}
\value{
Returns the dataset with the samples aggregated.
}
\description{
Aggregate samples according to an aggregate function like mean, median, etc.
This can be used to merge replicates.
}
\examples{
library(specmine.datasets)
data(propolis)
dataset = aggregate_samples(propolis, as.integer(propolis$metadata$seasons), "mean")

}
\keyword{aggregation}
\keyword{sample}
