% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{aov_all_vars}
\alias{aov_all_vars}
\title{Analysis of variance}
\usage{
aov_all_vars(
  dataset,
  column.class,
  doTukey = TRUE,
  write.file = FALSE,
  file.out = "anova-res.csv"
)
}
\arguments{
\item{dataset}{Dataset to analyze.}

\item{column.class}{Metadata column used to define groups.}

\item{doTukey}{Logical. If TRUE, also performs TukeyHSD post-hoc test.}

\item{write.file}{Logical. If TRUE, writes results to file.}

\item{file.out}{Output file name.}
}
\value{
A data frame with ANOVA results for all variables.
}
\description{
Performs one-way ANOVA for all variables in a dataset.
}
\examples{
## Example of ANOVA with TukeyHSD
if (requireNamespace("specmine.datasets", quietly = TRUE)) {
  data(propolis, package = "specmine.datasets")
  propolis_proc = missingvalues_imputation(propolis)
  propolis_proc = flat_pattern_filter(propolis_proc, "iqr", by.percent = TRUE, red.value = 75)
  result = aov_all_vars(propolis_proc, "seasons", doTukey = FALSE)
}

}
\keyword{anova}
\keyword{tukey}
