% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{apply_by_group}
\alias{apply_by_group}
\title{Apply by group}
\usage{
apply_by_group(dataset, fn.to.apply, metadata.var, var.value)
}
\arguments{
\item{dataset}{Dataset to analyze.}

\item{fn.to.apply}{Function to apply.}

\item{metadata.var}{Metadata variable used to define the group.}

\item{var.value}{Value or values of the metadata variable to select.}
}
\value{
Result of applying the function by variable for the selected group.
}
\description{
Applies a function to the variables of samples belonging to a given group.
}
\examples{
## Example of applying a function to a group
library(specmine.datasets)
data(cachexia)
apply.group.result = apply_by_group(cachexia, mean, "Muscle.loss", "control")

}
\keyword{apply}
\keyword{group}
