% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{apply_by_groups}
\alias{apply_by_groups}
\title{Apply by groups}
\usage{
apply_by_groups(
  dataset,
  metadata.var,
  fn.to.apply = "mean",
  variables = NULL,
  variable.bounds = NULL
)
}
\arguments{
\item{dataset}{Dataset to analyze.}

\item{metadata.var}{Metadata variable used to define groups.}

\item{fn.to.apply}{Function to apply.}

\item{variables}{Variables to include.}

\item{variable.bounds}{Optional numeric bounds for variables.}
}
\value{
A table of grouped results.
}
\description{
Applies a function to groups defined by a metadata variable.
}
\examples{
if (requireNamespace("specmine.datasets", quietly = TRUE)) {
  data(cachexia, package = "specmine.datasets")
  apply.groups.result = apply_by_groups(cachexia, "Muscle.loss", mean)
}

}
\keyword{apply}
\keyword{groups}
