% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{baseline_correction}
\alias{baseline_correction}
\title{Auto-exported function: baseline_correction}
\usage{
baseline_correction(dataset, method = "modpolyfit", ...)
}
\description{
Auto-exported function: baseline_correction
}
\keyword{internal}
