% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHWAY_ANALYSIS.R
\name{convert_chebi_to_kegg}
\alias{convert_chebi_to_kegg}
\title{Get kegg codes from chebi codes:}
\usage{
convert_chebi_to_kegg(chebi_codes)
}
\description{
Auto-exported function: convert_chebi_to_kegg
}
\keyword{internal}
