% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_chemospec.R
\name{convert_from_chemospec}
\alias{convert_from_chemospec}
\title{Auto-exported function: convert_from_chemospec}
\usage{
convert_from_chemospec(csobj, type = "undefined", description = "")
}
\description{
Auto-exported function: convert_from_chemospec
}
\keyword{internal}
