% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHWAY_ANALYSIS.R
\name{convert_hmdb_to_kegg}
\alias{convert_hmdb_to_kegg}
\title{Get kegg codes from hmdb codes:}
\usage{
convert_hmdb_to_kegg(hmdb_codes)
}
\arguments{
\item{hmdb_codes}{TODO.}
}
\description{
Get kegg codes from hmdb codes:
}
