\name{convert_keggpathway_2_reactiongraph}
\alias{convert_keggpathway_2_reactiongraph}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Convert KEGGPathway object to graph object.
}
\description{
Converts KEGGPathway object into a graph object.
}
\usage{
convert_keggpathway_2_reactiongraph(pathObj)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{pathObj}{KEGGPathway object.}
}
\value{
Kegg reaction graph.
}
