% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHWAY_ANALYSIS.R
\name{convert_keggpathway_2_reactiongraph}
\alias{convert_keggpathway_2_reactiongraph}
\title{Convert KEGGPathway object to graph object}
\usage{
convert_keggpathway_2_reactiongraph(pathObj)
}
\arguments{
\item{pathObj}{TODO.}
}
\description{
Convert KEGGPathway object to graph object
}
