% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHWAY_ANALYSIS.R
\name{convert_multiple_spcmnm_to_kegg}
\alias{convert_multiple_spcmnm_to_kegg}
\title{Get kegg codes from spcmnm codes:}
\usage{
convert_multiple_spcmnm_to_kegg(spcmnm_codes)
}
\arguments{
\item{spcmnm_codes}{TODO.}
}
\description{
Get kegg codes from spcmnm codes:
}
