% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{count_missing_values}
\alias{count_missing_values}
\title{Count missing values}
\usage{
count_missing_values(dataset)
}
\arguments{
\item{dataset}{Dataset to inspect.}
}
\value{
Total missing values.
}
\description{
Returns the total number of missing values in the dataset.
}
