% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{create_dataset}
\alias{create_dataset}
\title{Create dataset}
\usage{
create_dataset(
  datamatrix,
  type = "undefined",
  metadata = NULL,
  description = "",
  sample.names = NULL,
  x.axis.values = NULL,
  label.x = NULL,
  label.values = NULL,
  xSet = NULL
)
}
\arguments{
\item{datamatrix}{Matrix with numerical data; rows are assumed to be
variables and columns are assumed to be samples.}

\item{type}{Type of data, such as `"nmr-spectra"`, `"nmr-peaks"`,
`"ir-spectra"`, `"uvv-spectra"`, `"concentrations"`, or `"undefined"`.}

\item{metadata}{Optional metadata as a data frame or matrix.}

\item{description}{Dataset description.}

\item{sample.names}{Optional sample names.}

\item{x.axis.values}{Optional x-axis values.}

\item{label.x}{Optional x-axis label.}

\item{label.values}{Optional values label.}

\item{xSet}{Optional xSet object.}
}
\value{
A dataset object as a list.
}
\description{
Creates a dataset from existing objects.
}
