% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHWAY_ANALYSIS.R
\name{create_pathway_with_reactions}
\alias{create_pathway_with_reactions}
\title{Creates the pathway, with reactions included in the nodes}
\usage{
create_pathway_with_reactions(
  path,
  path.name,
  identified_cpds,
  nodeNames = "kegg",
  nodeTooltip = FALSE,
  map.zoom = FALSE,
  map.layout = "preset",
  map.width = NULL,
  map.height = NULL
)
}
\arguments{
\item{path}{TODO.}

\item{path.name}{TODO.}

\item{identified_cpds}{TODO.}

\item{nodeNames}{TODO.}

\item{nodeTooltip}{TODO.}

\item{map.zoom}{TODO.}

\item{map.layout}{TODO.}

\item{map.width}{TODO.}

\item{map.height}{TODO.}
}
\description{
Creates the pathway, with reactions included in the nodes
}
