% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{feature_selection}
\alias{feature_selection}
\title{Auto-exported function: feature_selection}
\usage{
feature_selection(
  dataset,
  column.class,
  method = "rfe",
  functions,
  validation = "cv",
  repeats = 5,
  number = 10,
  subsets = 2^(2:4)
)
}
\description{
Auto-exported function: feature_selection
}
\keyword{internal}
