% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMR_metabolite_identification.R
\name{find_corr}
\alias{find_corr}
\title{FIND THE BEST CORRELATION VALUE FOR THE CLUSTERING OF PEAKS}
\usage{
find_corr(dataset_orig, CMETH = "pearson")
}
\arguments{
\item{dataset_orig}{List representing the dataset from an nmr peaks metabolomics experiment.}

\item{CMETH}{Correlation method used to cluster the variables. Defaults to "pearson".}
}
\value{
Value of the optimal correlation.
}
\description{
Takes a dataset and calculates the optimum correlation value, which leads to the maximum
number of clusters of the variables.
}
