% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters_flat.R
\name{flat_pattern_filter}
\alias{flat_pattern_filter}
\title{Flat pattern filter}
\usage{
flat_pattern_filter(
  dataset,
  filter.function = "iqr",
  by.percent = TRUE,
  by.threshold = FALSE,
  red.value = 0
)
}
\arguments{
\item{dataset}{Dataset to filter.}

\item{filter.function}{Filtering function to use. One of
`"iqr"`, `"rsd"`, `"rnsd"`, `"sd"`, `"mad"`, `"mean"`, or `"median"`.}

\item{by.percent}{Logical. If `TRUE`, the number of variables to filter
will be defined as a percentage of the number of variables in the dataset;
percentage is given by `red.value`.}

\item{by.threshold}{Logical. If `TRUE`, filtering selects variables where
the filtering function is above or equal to a threshold.}

\item{red.value}{Reduction value. If `by.percent = TRUE`, this is the
percentage of variables to remove, or `"auto"` for automatic calculation.
If `by.threshold = TRUE`, this is the minimum value needed to keep
the variable.}
}
\value{
Filtered dataset.
}
\description{
Performs a flat pattern filter over the dataset.
}
\examples{
if (requireNamespace("specmine.datasets", quietly = TRUE)) {
  data(propolis, package = "specmine.datasets")
  propolis_proc = missingvalues_imputation(propolis)
  propolis_proc = flat_pattern_filter(propolis_proc, "iqr",
    by.percent = TRUE, red.value = 75)
}

}
