\name{get_cpd_names}
\alias{get_cpd_names}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get the names of the compounds that correspond to the kegg codes given
}
\usage{
get_cpds_names(kegg_codes)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{kegg_codes}{
 Vector with kegg codes.
}
}
\value{
Named vector with the names of the compounds. The bames of the vector are the compounds names and the vector elements the kegg codes.
}
\examples{

}
