\name{get_MetabolitePath}
\alias{get_MetabolitePath}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 Returns an object of KEGGPathway of the pathway especified in pathcode.
}
\description{
Returns an object of KEGGPathway of the pathway specified in pathcode (e.g. "hsa00010").
}
\usage{
get_MetabolitePath(pathcode)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{pathcode}{Pathway code of the path wanted.}
}
\value{
KEGGPathway object.
}
