% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHWAY_ANALYSIS.R
\name{get_MetabolitePath}
\alias{get_MetabolitePath}
\title{Returns an object of KEGGPathway of the pathway especified in pathcode}
\usage{
get_MetabolitePath(pathcode)
}
\arguments{
\item{pathcode}{TODO.}
}
\description{
Returns an object of KEGGPathway of the pathway especified in pathcode
}
