\name{get_metabPaths_org}
\alias{get_metabPaths_org}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Get the metabolic pathways present in given organism.}
\description{
Get vector with paths numbers that occur in the given organism, named with the full paths names.
}
\usage{
get_metabPaths_org(org_code)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{org_code}{
Organism code. The correct code for an organism can be consulted using function \code{\link{get_OrganismsCodes}}.
}
}
