% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHWAY_ANALYSIS.R
\name{get_metabPaths_org}
\alias{get_metabPaths_org}
\title{Get vector with paths numbers that occur in the given organism, named with the full path name:}
\usage{
get_metabPaths_org(org_code)
}
\arguments{
\item{org_code}{TODO.}
}
\description{
Get vector with paths numbers that occur in the given organism, named with the full path name:
}
