\name{get_OrganismsCodes}
\alias{get_OrganismsCodes}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Get all organisms in KEGG.
}
\description{
Get code, t number, full name and phylogeny of all organisms in KEGG.
}
\usage{
get_OrganismsCodes()
}
\value{
Data frame with t number, organism code, full name and phylogeny for each kegg organism.
}
\examples{
\donttest{
get_OrganismsCodes()
}
}
