% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHWAY_ANALYSIS.R
\name{get_OrganismsCodes}
\alias{get_OrganismsCodes}
\title{Get code, t number, full name and phylogeny of all organisms in KEGG:}
\usage{
get_OrganismsCodes()
}
\description{
Get code, t number, full name and phylogeny of all organisms in KEGG:
}
