% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHWAY_ANALYSIS.R
\name{get_paths_with_cpds_org}
\alias{get_paths_with_cpds_org}
\title{Get only the paths of the organism that contain given compounds:}
\usage{
get_paths_with_cpds_org(organism_code, compounds, full.result = TRUE)
}
\arguments{
\item{organism_code}{TODO.}

\item{compounds}{TODO.}

\item{full.result}{TODO.}
}
\description{
Get only the paths of the organism that contain given compounds:
}
