% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dx.R
\name{get_samples_names_dx}
\alias{get_samples_names_dx}
\title{Auto-exported function: get_samples_names_dx}
\usage{
get_samples_names_dx(foldername)
}
\description{
Auto-exported function: get_samples_names_dx
}
\keyword{internal}
