% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{get_x_label}
\alias{get_x_label}
\title{Get x label}
\usage{
get_x_label(dataset)
}
\arguments{
\item{dataset}{Dataset object.}
}
\value{
Character string with the x-axis label.
}
\description{
Returns the x-axis label associated with the dataset.
}
