% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{get_x_values_as_text}
\alias{get_x_values_as_text}
\title{Get x values as text}
\usage{
get_x_values_as_text(dataset)
}
\arguments{
\item{dataset}{Dataset object.}
}
\value{
Character vector with x values.
}
\description{
Returns the x values of a dataset as text.
}
