% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{impute_nas_knn}
\alias{impute_nas_knn}
\title{Impute missing values with kNN}
\usage{
impute_nas_knn(dataset, k = 10, ...)
}
\arguments{
\item{dataset}{Dataset to modify.}

\item{k}{Number of neighbors.}

\item{...}{Additional arguments passed to `impute::impute.knn()`.}
}
\value{
Modified dataset.
}
\description{
Impute missing values with kNN
}
