% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{impute_nas_mean}
\alias{impute_nas_mean}
\title{Impute missing values with mean}
\usage{
impute_nas_mean(dataset)
}
\arguments{
\item{dataset}{Dataset to modify.}
}
\value{
Modified dataset.
}
\description{
Impute missing values with mean
}
