% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{impute_nas_median}
\alias{impute_nas_median}
\title{Impute missing values with median}
\usage{
impute_nas_median(dataset)
}
\arguments{
\item{dataset}{Dataset to modify.}
}
\value{
Modified dataset.
}
\description{
Impute missing values with median
}
