% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{impute_nas_value}
\alias{impute_nas_value}
\title{Impute missing values with a constant}
\usage{
impute_nas_value(dataset, value)
}
\arguments{
\item{dataset}{Dataset to modify.}

\item{value}{Replacement value.}
}
\value{
Modified dataset.
}
\description{
Impute missing values with a constant
}
