% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{merge_data_metadata}
\alias{merge_data_metadata}
\title{Merge data and metadata}
\usage{
merge_data_metadata(
  dataset,
  samples = NULL,
  metadata.vars = NULL,
  x.values = NULL,
  by.index = FALSE
)
}
\arguments{
\item{dataset}{Dataset to merge.}

\item{samples}{Samples to keep.}

\item{metadata.vars}{Metadata variables to keep.}

\item{x.values}{X values to keep.}

\item{by.index}{Logical. If TRUE, x.values are indexes.}
}
\value{
A data frame with merged data and metadata.
}
\description{
Merges data and metadata into a data frame.
}
