\name{metabolights_studies_list}
\alias{metabolights_studies_list}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
List the study IDs available in the MetaboLights database.
}
\description{
Gives the IDs of the studies available in the MetaboLights database.
}
\usage{
metabolights_studies_list()
}
\value{
Vector with the different study IDs available in the MetaboLights database.
}
\references{
MetaboLights database: https://www.ebi.ac.uk/metabolights/
}
\examples{
## metabolights_studies_list()
}
