% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{missingvalues_imputation}
\alias{missingvalues_imputation}
\title{Missing values imputation}
\usage{
missingvalues_imputation(dataset, method = "value", value = 5e-04, k = 5)
}
\arguments{
\item{dataset}{Dataset to inspect.}

\item{method}{Imputation method: "value", "mean", "median", "knn", or "linapprox".}

\item{value}{If method = "value", the value that will replace NAs.}

\item{k}{If method = "knn", the number of neighbors.}
}
\value{
Dataset with imputed missing values.
}
\description{
Impute missing values in a dataset using different methods.
}
\examples{
library(specmine.datasets)
data(propolis)
propolis_proc = missingvalues_imputation(propolis, method = "value", value = 0.0005)

}
