\name{multiClassSummary}
\alias{multiClassSummary}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Multi Class Summary
}
\description{
  Summary function for caret to compute AUC.
}
\usage{
multiClassSummary(data, lev = NULL, model = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{
	data parameter.
}
  \item{lev}{
	lev parameter.
}
  \item{model}{
	model parameter.
}
}

\references{
www.r-bloggers.com/error-metrics-for-multi-class-problems-in-r-beyond-accuracy-and-kappa/
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ caret }
\keyword{ summaryfunction }% __ONLY ONE__ keyword per line
