% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PATHWAY_ANALYSIS.R
\name{pathway_analysis}
\alias{pathway_analysis}
\title{Creates the pathway wanted. If any of the given compounds is present in the pathway, it is coloured differently.}
\usage{
pathway_analysis(
  compounds,
  pathway,
  nodeNames = "kegg",
  nodeTooltip = FALSE,
  map.zoom = FALSE,
  map.layout = "preset",
  map.width = NULL,
  map.height = NULL
)
}
\arguments{
\item{compounds}{TODO.}

\item{pathway}{TODO.}

\item{nodeNames}{TODO.}

\item{nodeTooltip}{TODO.}

\item{map.zoom}{TODO.}

\item{map.layout}{TODO.}

\item{map.width}{TODO.}

\item{map.height}{TODO.}
}
\description{
Creates the pathway wanted. If any of the given compounds is present in the pathway, it is coloured differently.
}
