% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning.R
\name{pca_plot_3d}
\alias{pca_plot_3d}
\title{Auto-exported function: pca_plot_3d}
\usage{
pca_plot_3d(
  dataset,
  model,
  var.class,
  pcas = 1:3,
  colors = NULL,
  legend.place = "topright",
  ...
)
}
\description{
Auto-exported function: pca_plot_3d
}
\keyword{internal}
