% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_scoresplot2D}
\alias{pca_scoresplot2D}
\title{PCA 2D scores plot}
\usage{
pca_scoresplot2D(
  dataset,
  pca.result,
  column.class = NULL,
  pcas = c(1, 2),
  labels = FALSE,
  ellipses = FALSE,
  bw = FALSE,
  pallette = 2,
  leg.pos = "right",
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{dataset}{Dataset used in the PCA.}

\item{pca.result}{PCA result object.}

\item{column.class}{Optional metadata column used for colouring groups.}

\item{pcas}{Principal components to plot.}

\item{labels}{Logical indicating whether sample labels should be shown.}

\item{ellipses}{Logical indicating whether group ellipses should be drawn.}

\item{bw}{Logical indicating whether a black-and-white style should be used.}

\item{pallette}{Brewer palette identifier.}

\item{leg.pos}{Legend position.}

\item{xlim}{Optional x-axis limits.}

\item{ylim}{Optional y-axis limits.}
}
\value{
A \code{ggplot} object.
}
\description{
Creates a 2D PCA scores plot.
}
\examples{
\donttest{
datamat <- matrix(
  rnorm(20),
  nrow = 4,
  dimnames = list(paste0("x", 1:4), paste0("s", 1:5))
)
metadata <- data.frame(class = factor(c("A", "A", "B", "B", "A")))
dataset <- list(data = datamat, metadata = metadata)
pca_res <- pca_analysis_dataset(dataset)
pca_scoresplot2D(dataset, pca_res, column.class = "class")
}

}
