% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_data.R
\name{read_data_csv}
\alias{read_data_csv}
\title{Auto-exported function: read_data_csv}
\usage{
read_data_csv(
  filename,
  format = "row",
  header.col = TRUE,
  header.row = TRUE,
  sep = ","
)
}
\description{
Auto-exported function: read_data_csv
}
\keyword{internal}
