% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_functions.R
\name{read_ms_spectra}
\alias{read_ms_spectra}
\title{Auto-exported function: read_ms_spectra}
\usage{
read_ms_spectra(
  folder.name,
  type = "undefined",
  filename.meta = NULL,
  description = "",
  prof.method = "bin",
  fwhm = 30,
  bw = 30,
  intvalue = "into",
  header.col.meta = TRUE,
  header.row.meta = TRUE,
  sep.meta = ","
)
}
\description{
Auto-exported function: read_ms_spectra
}
\keyword{internal}
