% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{remove_data}
\alias{remove_data}
\title{Remove data}
\usage{
remove_data(
  dataset,
  data.to.remove,
  type = "sample",
  by.index = FALSE,
  rebuild.factors = TRUE
)
}
\arguments{
\item{dataset}{Dataset to modify.}

\item{data.to.remove}{Data to remove.}

\item{type}{One of "sample", "data", or "metadata".}

\item{by.index}{Logical. If TRUE, data.to.remove has indexes.}

\item{rebuild.factors}{Logical. If TRUE, rebuild factors in metadata.}
}
\value{
Modified dataset.
}
\description{
Removes selected samples, data variables, or metadata variables from a dataset.
}
