% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{remove_data_variables}
\alias{remove_data_variables}
\title{Remove data variables}
\usage{
remove_data_variables(dataset, variables.to.remove, by.index = FALSE)
}
\arguments{
\item{dataset}{Dataset to modify.}

\item{variables.to.remove}{Variables to remove.}

\item{by.index}{Logical. If TRUE, variables.to.remove are indexes.}
}
\value{
Modified dataset.
}
\description{
Removes x variables from a dataset.
}
