% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{remove_metadata_variables}
\alias{remove_metadata_variables}
\title{Remove metadata variables}
\usage{
remove_metadata_variables(dataset, variables.to.remove)
}
\arguments{
\item{dataset}{Dataset to modify.}

\item{variables.to.remove}{Metadata variables to remove.}
}
\value{
Modified dataset.
}
\description{
Removes metadata variables from a dataset.
}
