% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{remove_samples}
\alias{remove_samples}
\title{Remove samples}
\usage{
remove_samples(dataset, samples.to.remove, rebuild.factors = TRUE)
}
\arguments{
\item{dataset}{Dataset to modify.}

\item{samples.to.remove}{Samples to remove.}

\item{rebuild.factors}{Logical. If TRUE, rebuild factors in metadata.}
}
\value{
Modified dataset.
}
\description{
Removes samples from a dataset.
}
