% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{remove_samples_by_na_metadata}
\alias{remove_samples_by_na_metadata}
\title{Remove samples by NA metadata}
\usage{
remove_samples_by_na_metadata(dataset, metadata.var)
}
\arguments{
\item{dataset}{Dataset to modify.}

\item{metadata.var}{Metadata variable name.}
}
\value{
Modified dataset.
}
\description{
Removes samples with NA in a given metadata variable.
}
