% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{remove_variables_by_nas}
\alias{remove_variables_by_nas}
\title{Remove variables by NAs}
\usage{
remove_variables_by_nas(dataset, max.nas = 0, by.percent = FALSE)
}
\arguments{
\item{dataset}{Dataset to modify.}

\item{max.nas}{Maximum number of missing values allowed.}

\item{by.percent}{Logical. If TRUE, max.nas is treated as a percentage.}
}
\value{
Modified dataset.
}
\description{
Removes variables with too many missing values.
}
