% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{remove_x_values_by_interval}
\alias{remove_x_values_by_interval}
\title{Remove x values by interval}
\usage{
remove_x_values_by_interval(dataset, min.value, max.value)
}
\arguments{
\item{dataset}{Dataset to modify.}

\item{min.value}{Minimum x value.}

\item{max.value}{Maximum x value.}
}
\value{
Modified dataset.
}
\description{
Removes x values inside an interval.
}
