% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R, R/specmine-package.R
\docType{package}
\name{specmine-package}
\alias{specmine-package}
\title{specmine: Metabolomics and Spectral Data Analysis and Mining}
\description{
Provides methods for metabolomics and spectral data analysis, including data import, preprocessing, visualization, univariate and multivariate analysis, machine learning, feature selection, and pathway analysis. The package supports analytical workflows for different data types used in metabolomics and spectroscopy. Case studies and additional documentation are available in the project repository.

Package-level imports for specmine.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PedroFontao/specmine}
  \item Report bugs at \url{https://github.com/PedroFontao/specmine/issues}
}

}
\author{
\strong{Maintainer}: Pedro Fontão \email{pedrofontao812004@gmail.com}

Authors:
\itemize{
  \item Pedro Fontão \email{pedrofontao812004@gmail.com}
  \item Christopher Costa \email{chrisbcl@hotmail.com}
  \item Marcelo Maraschin \email{mtocsy@gmail.com}
  \item Miguel Rocha \email{mrocha@di.uminho.pt}
  \item Sara Cardoso \email{saracardoso501@gmail.com}
  \item Telma Afonso \email{telma.afonso94@gmail.com}
  \item Bruno Pereira \email{pereirinha_bp@hotmail.com}
}

Other contributors:
\itemize{
  \item C. Beleites [copyright holder]
  \item Jie Hao [copyright holder]
}

}
\keyword{internal}
