% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_options.R
\docType{data}
\name{spectra_options}
\alias{spectra_options}
\title{Spectra options}
\format{
A named list of options.
}
\usage{
spectra_options
}
\description{
Default options used by spectra-related utilities in the package.
}
\keyword{datasets}
