% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{subset_by_samples_and_xvalues}
\alias{subset_by_samples_and_xvalues}
\title{Subset by samples and x values}
\usage{
subset_by_samples_and_xvalues(
  dataset,
  samples,
  variables = NULL,
  by.index = FALSE,
  variable.bounds = NULL,
  rebuild.factors = TRUE
)
}
\arguments{
\item{dataset}{Dataset to subset.}

\item{samples}{Sample indexes.}

\item{variables}{Variables to keep.}

\item{by.index}{Logical. If TRUE, variables are interpreted as indexes.}

\item{variable.bounds}{Optional numeric bounds for x values.}

\item{rebuild.factors}{If TRUE, rebuild factors in metadata.}
}
\value{
Subset dataset.
}
\description{
Selects samples and x values simultaneously.
}
