% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{subset_metadata}
\alias{subset_metadata}
\title{Subset metadata}
\usage{
subset_metadata(dataset, variables)
}
\arguments{
\item{dataset}{Dataset to subset.}

\item{variables}{Metadata variables to keep.}
}
\value{
Dataset with selected metadata variables.
}
\description{
Selects metadata variables to keep.
}
