% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{subset_random_samples}
\alias{subset_random_samples}
\title{Subset random samples}
\usage{
subset_random_samples(dataset, nsamples)
}
\arguments{
\item{dataset}{Dataset to subset.}

\item{nsamples}{Number of samples to select.}
}
\value{
Dataset with selected random samples.
}
\description{
Selects a random subset of samples from the dataset.
}
