% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{subset_samples}
\alias{subset_samples}
\title{Subset samples}
\usage{
subset_samples(dataset, samples, rebuild.factors = TRUE)
}
\arguments{
\item{dataset}{Dataset to subset.}

\item{samples}{Vector with indexes or names of the samples to select.}

\item{rebuild.factors}{If TRUE, rebuild factors in metadata.}
}
\value{
Dataset with selected samples.
}
\description{
Returns a dataset with a selected set of samples.
}
