% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{subset_samples_by_metadata_values}
\alias{subset_samples_by_metadata_values}
\title{Subset samples by metadata values}
\usage{
subset_samples_by_metadata_values(dataset, metadata.varname, values)
}
\arguments{
\item{dataset}{Dataset to subset.}

\item{metadata.varname}{Metadata variable name.}

\item{values}{Values to keep.}
}
\value{
Dataset with selected samples.
}
\description{
Selects a set of samples by the value of a metadata variable.
}
