% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{subset_x_values}
\alias{subset_x_values}
\title{Subset x values}
\usage{
subset_x_values(dataset, variables, by.index = FALSE)
}
\arguments{
\item{dataset}{Dataset to subset.}

\item{variables}{Variables to keep.}

\item{by.index}{Logical. If TRUE, variables are interpreted as row indexes.}
}
\value{
Dataset with selected x values.
}
\description{
Selects rows of the dataset by x values or by row index.
}
