% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dataset.R
\name{subset_x_values_by_interval}
\alias{subset_x_values_by_interval}
\title{Subset x values by interval}
\usage{
subset_x_values_by_interval(dataset, min.value, max.value)
}
\arguments{
\item{dataset}{Dataset to subset.}

\item{min.value}{Minimum x value.}

\item{max.value}{Maximum x value.}
}
\value{
Dataset with selected x values.
}
\description{
Selects x values within a numeric interval.
}
